refactor: replace supervisord with s6
This change looks to drop the python dependency in the base image and replaces supervisord with [s6](https://github.com/just-containers/s6-overlay).

If we could get this branch pushed to the `s6` tag on Docker it will allow us to test this with the downstream containers.

I will update the README.md if it's decided that we will adopt s6 over supervisord.